pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-backtesting-engine"
version = "1.0.0"
description = "Professional-grade multi-strategy backtesting engine with portfolio optimization"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Quant Trading System"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # Core Data Science
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",

    # Financial Data
    "yfinance>=0.2.28",
    "alpha-vantage>=2.3.1",
    "ccxt>=4.0.0",
    "pandas-datareader>=0.10.0",

    # Technical Analysis
    "pandas-ta>=0.3.14b",

    # Storage & Performance
    "tables>=3.8.0",
    "pyarrow>=12.0.0",
    "fastparquet>=2023.7.0",
    "h5py>=3.9.0",

    # Portfolio Optimization
    "cvxpy>=1.3.0",
    "PyPortfolioOpt>=1.5.5",

    # Machine Learning
    "xgboost>=1.7.0",
    "lightgbm>=4.0.0",
    "catboost>=1.2.0",
    "statsmodels>=0.14.0",
    "hmmlearn>=0.3.0",

    # Reinforcement Learning
    "stable-baselines3>=2.1.0",
    "gymnasium>=0.29.0",
    "torch>=2.0.0",

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",
    "dash>=2.11.0",
    "streamlit>=1.25.0",
    "altair>=5.0.0",

    # Report Generation
    "reportlab>=4.0.0",
    "Pillow>=10.0.0",
    "fpdf2>=2.7.0",

    # Utils
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "tqdm>=4.65.0",
    "joblib>=1.3.0",
    "numba>=0.57.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "mypy>=1.4.0",
    "types-requests>=2.31.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"
